Modify/Refactor Cluster (commands) tests
