+++ src/test/java/redis/clients/jedis/SSLACLJedisClusterTest.java
@@ -18,7 +18,7 @@ import redis.clients.jedis.exceptions.JedisClusterOperationException;
 import redis.clients.jedis.SSLJedisTest.BasicHostnameVerifier;
 import redis.clients.jedis.util.RedisVersionUtil;
 
-public class SSLACLJedisClusterTest extends JedisClusterTest {
+public class SSLACLJedisClusterTest extends JedisClusterTestBase {
 
   private static final int DEFAULT_REDIRECTIONS = 5;
   private static final ConnectionPoolConfig DEFAULT_POOL_CONFIG = new ConnectionPoolConfig();
